backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# LangChain ecosystem
langchain==0.1.0
langserve==0.0.30
langgraph==0.0.20
langchain-community==0.0.10

# EURI AI
euriai

# Vector databases
chromadb==0.4.18
faiss-cpu==1.7.4

# PDF processing
PyMuPDF==1.23.8
pdfplumber==0.10.3
camelot-py==0.11.0
pytesseract==0.3.10
Pillow==10.1.0
opencv-python==4.8.1.78

# Data processing
pandas==2.1.4
numpy==1.25.2
plotly==5.17.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
psutil==5.9.6

# Caching
redis==5.0.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
requests==2.31.0
aiofiles==23.2.1
tenacity==8.5.0

# Logging and monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2